backend/requirements.txt
fastapi
uvicorn
google-cloud-vision
google-auth
requests
python-multipart
